re-index attacks (keep id 0 as benign)
