+++ pkg/src/attacks/attack10.rs
@@ -4,58 +4,47 @@ use crate::sys::{
 };
 
 #[derive(Clone, Debug)]
-pub struct CollisionAttackAgainstTheBus {
-    pub nwords_inj: u32,
-    pub started: u128,
+pub struct CommandInvalidationAttack {
+    pub attack_times: Vec<u128>,
     pub success: bool,
+    pub target: u8,         // the target RT
+    pub target_found: bool, // target found in traffic
 }
 
-impl CollisionAttackAgainstTheBus {
+impl CommandInvalidationAttack {
     pub fn inject(&mut self, d: &mut Device) {
-        if self.started == 0 {
-            self.started = d.clock.elapsed().as_nanos();
-            self.success = true;
-        }
-        for i in 0..self.nwords_inj {
-            let w = Word::new_data(i as u32);
-            d.log(
-                WRD_EMPTY,
-                ErrMsg::MsgAttk(format!("Sent Fake Data {} ", w).to_string()),
-            );
-            d.write(w);
-        }
+        self.attack_times.push(d.clock.elapsed().as_nanos());
+        let dword_count = 31;    // Maximum number of words.  This will mean the receipient ignores the next 31 messages
+        let tr = 0;              // 1: transmit, 0: receive.  We want to receive because it will sit and wait rather than responding to the BC.
+        let w = Word::new_cmd(self.target, dword_count, tr);
+        d.log(
+            WRD_EMPTY,
+            ErrMsg::MsgAttk(format!("Attacker>> Injecting fake command on RT{}", w).to_string()),
+        );
+        d.write(w);
+        self.success = true;
     }
 }
 
-impl EventHandler for CollisionAttackAgainstTheBus {
+impl EventHandler for CommandInvalidationAttack {
     fn on_cmd(&mut self, d: &mut Device, w: &mut Word) {
-        d.log(
-            *w,
-            ErrMsg::MsgAttk("Jamming launched (after cmd)".to_string()),
-        );
-        self.inject(d);
+        // This function replaces "find_RT_tcmd" from Michael's code
+        // We cannot use on_cmd_trx here because that only fires after on_cmd verifies that the address is correct.
+        let destination = w.address();
+        if destination == self.target && self.target_found==false && w.tr() == 1 {
+            d.log(
+                *w, 
+                ErrMsg::MsgAttk(format!("Attacker>> Target detected(RT{})", self.target).to_string()),
+            );
+            self.target_found = true;
+            self.inject(d);
+        }
         self.default_on_cmd(d, w);
-    }
-    fn on_dat(&mut self, d: &mut Device, w: &mut Word) {
-        d.log(
-            *w,
-            ErrMsg::MsgAttk("Jamming launched (after data)".to_string()),
-        );
-        self.inject(d);
-        self.default_on_dat(d, w);
-    }
-    fn on_sts(&mut self, d: &mut Device, w: &mut Word) {
-        d.log(
-            *w,
-            ErrMsg::MsgAttk("Jamming launched (after status)".to_string()),
-        );
-        self.inject(d);
-        self.default_on_dat(d, w);
     }
 }
 
 #[allow(dead_code)]
-pub fn test_attack0() {
+pub fn test_attack10() {
     // let mut delays_single = Vec::new();
     let n_devices = 8;
     // normal device has 4ns delays (while attacker has zero)
@@ -91,10 +80,11 @@ pub fn test_attack0() {
             proto: 0,
         },
         // control device-level response
-        handler: CollisionAttackAgainstTheBus {
-            nwords_inj: 5,
-            started: 0,
+        handler: CommandInvalidationAttack {
+            attack_times: Vec::new(),
             success: false,
+            target: 4, // attacking RT address @4
+            target_found: false,
         },
     };
