+++ pkg/src/attacks/mod.rs
@@ -1,5 +1,5 @@
-pub mod attack0;
 pub mod attack1;
+pub mod attack10;
 pub mod attack2;
 pub mod attack3;
 pub mod attack4;
@@ -9,5 +9,8 @@ pub mod attack7;
 pub mod attack8;
 pub mod attack9;
 
-pub use {attack0::test_attack0, attack1::test_attack1, attack2::test_attack2, attack3::test_attack3, attack4::test_attack4, 
-    attack5::test_attack5, attack6::test_attack6, attack7::test_attack7, attack8::test_attack8, attack9::test_attack9};
+pub use {
+    attack1::test_attack1, attack10::test_attack10, attack2::test_attack2, attack3::test_attack3,
+    attack4::test_attack4, attack5::test_attack5, attack6::test_attack6, attack7::test_attack7,
+    attack8::test_attack8, attack9::test_attack9,
+};
