+++ pkg/src/main.rs
@@ -1,8 +1,10 @@
 mod attacks;
 mod sys;
 #[allow(unused_imports)]
-use attacks::{test_attack0, test_attack1, test_attack2, test_attack3,
-        test_attack4, test_attack5, test_attack6, test_attack7, test_attack8, test_attack9};
+use attacks::{
+    test_attack1, test_attack10, test_attack2, test_attack3, test_attack4, test_attack5,
+    test_attack6, test_attack7, test_attack8, test_attack9,
+};
 #[allow(unused_imports)]
 use sys::test_default;
